smartContract/requirements.txt
algorand-sdk
pyteal
beaker-pyteal